Added packages to clean the repo up a bit.
Abstracted all of the protocol specific stuff to two classes, subclasses to LinkLayerPort and implementors of TransportLayerAbstraction.
LinkLayerPort imposes some abstract methods on the subclasses which are meant to be enclose the protocol specific code.
TransportLayerAbstraction is essentially  wrapper for sockets to make Connection protocol independent. This had to be done by "hiding"
the UDP sockets and packets in a stream, but it works okay. The packets are intercepted in UDPConnectionMultiplexer and are appended to the appriopriate
Connection stream. Also added LabComm version control in the encoders for the connection. In Channel the version is still supplied locally.
There is a bug at the application level in the ping/pong test where one of them closes to quickly and the other does not receive the echo.

Started working on a TCP implementation as well.