New application callbacks and protocol handling
Added new application callbacks and changed the old ones a little to
be more similar to the ones in the c-version of firefly.

Fixed some protocol implmentation issues, mainly constants but also
restrict message handling. It should work with the c-version but it
hasn't been tested.

Implemented ordering of important data packets and important buffer in
channel so that only one important packet is "on the connection"
concurrently, any non-expected packets are discared.

Also some packet layout restructure.